requirements.txt
Django==4.2.2
djangorestframework==3.14.0
django-filter==23.2
markdown==3.4.3
django-cors-headers==4.1.0
django-model-utils
requests==2.31.0
Pillow==9.5.0
python-dotenv==1.0.0
drf-writable-nested
django-constance[database]
psycopg2-binary==2.9.6
django-ckeditor
python-magic-bin
djangorestframework-simplejwt==5.2.2